Navigtion: Use the ListView in the Navigation block inspector controls
use the hook from list view

remove prop drilling

remove prop drilling

Allow list view to scroll

add custom scrollbars on hover

update navigation block tests to account for using the list view

fix test

List View: Add a new prop to allow blocks in the inserter to be prioritized

add another prop to list view

Don't bother sorting if there are no proitizedInnerBlocks

Open LinkUI popover after appending link from Navigation side block inserter

The behavior when adding a link from the appender button on the OffCanvasEditor is to open a link control popover. This commit copies over several files from the OffCanvasEditor that was required for that to work in the ListView. The behavior is buggy, IMO, but it does match what is in trunk.

inverted colors for the scrollbar

Moved files related to navigation block outside of list-view.

Remove unused link-ui from list-view

Fix webpack runtime error

For some reason keeping both exports in the same file caused an error that PrivateListView was used before initialization

remove unneeded code to prioritize the blocks in the inserter

remove unused prop

move the inline function to a separate definition

remove unneccessary change
