+++ packages/block-library/src/navigation/edit/leaf-more-menu.js
@@ -29,9 +29,12 @@ function AddSubmenuItem( { block, onClose, expandedState, expand } ) {
 		useDispatch( blockEditorStore );
 
 	const clientId = block.clientId;
-	const isDisabled = ! BLOCKS_THAT_CAN_BE_CONVERTED_TO_SUBMENU.includes(
-		block.name
+	const blockName = useSelect(
+		( select ) => select( blockEditorStore ).getBlockName( clientId ),
+		[ clientId ]
 	);
+	const isDisabled =
+		! BLOCKS_THAT_CAN_BE_CONVERTED_TO_SUBMENU.includes( blockName );
 	return (
 		<MenuItem
 			icon={ addSubmenu }
