+++ packages/block-library/src/navigation/edit/menu-inspector-controls.js
@@ -13,6 +13,7 @@ import {
 	Spinner,
 } from '@wordpress/components';
 import { useSelect } from '@wordpress/data';
+import { useState, useEffect } from '@wordpress/element';
 import { __, sprintf } from '@wordpress/i18n';
 
 /**
@@ -23,9 +24,16 @@ import { unlock } from '../../private-apis';
 import DeletedNavigationWarning from './deleted-navigation-warning';
 import useNavigationMenu from '../use-navigation-menu';
 import LeafMoreMenu from './leaf-more-menu';
+import { updateAttributes } from '../../navigation-link/update-attributes';
+import { LinkUI } from '../../navigation-link/link-ui';
+import { useInsertedBlock } from '../../navigation-link/use-inserted-block';
 
 /* translators: %s: The name of a menu. */
 const actionLabel = __( "Switch to '%s'" );
+const BLOCKS_WITH_LINK_UI_SUPPORT = [
+	'core/navigation-link',
+	'core/navigation-submenu',
+];
 
 const MainContent = ( {
 	clientId,
@@ -34,7 +42,7 @@ const MainContent = ( {
 	isNavigationMenuMissing,
 	onCreateNew,
 } ) => {
-	const { OffCanvasEditor } = unlock( blockEditorPrivateApis );
+	const { PrivateListView } = unlock( blockEditorPrivateApis );
 
 	// Provide a hierarchy of clientIds for the given Navigation block (clientId).
 	// This is required else the list view will display the entire block tree.
@@ -45,6 +53,42 @@ const MainContent = ( {
 		},
 		[ clientId ]
 	);
+
+	const [ clientIdWithOpenLinkUI, setClientIdWithOpenLinkUI ] = useState();
+	const { lastInsertedBlockClientId } = useSelect( ( select ) => {
+		const { getLastInsertedBlocksClientIds } = unlock(
+			select( blockEditorStore )
+		);
+		const lastInsertedBlocksClientIds = getLastInsertedBlocksClientIds();
+		return {
+			lastInsertedBlockClientId:
+				lastInsertedBlocksClientIds && lastInsertedBlocksClientIds[ 0 ],
+		};
+	}, [] );
+
+	const {
+		insertedBlockAttributes,
+		insertedBlockName,
+		setInsertedBlockAttributes,
+	} = useInsertedBlock( lastInsertedBlockClientId );
+
+	const hasExistingLinkValue = insertedBlockAttributes?.url;
+
+	useEffect( () => {
+		if (
+			lastInsertedBlockClientId &&
+			BLOCKS_WITH_LINK_UI_SUPPORT?.includes( insertedBlockName ) &&
+			! hasExistingLinkValue // don't re-show the Link UI if the block already has a link value.
+		) {
+			setClientIdWithOpenLinkUI( lastInsertedBlockClientId );
+		}
+	}, [
+		lastInsertedBlockClientId,
+		clientId,
+		insertedBlockName,
+		hasExistingLinkValue,
+	] );
+
 	const { navigationMenu } = useNavigationMenu( currentMenuId );
 
 	if ( currentMenuId && isNavigationMenuMissing ) {
@@ -64,14 +108,41 @@ const MainContent = ( {
 		: __(
 				'You have not yet created any menus. Displaying a list of your Pages'
 		  );
+
+	const renderLinkUI = ( block ) => {
+		return (
+			clientIdWithOpenLinkUI === block.clientId && (
+				<LinkUI
+					clientId={ lastInsertedBlockClientId }
+					link={ insertedBlockAttributes }
+					onClose={ () => setClientIdWithOpenLinkUI( null ) }
+					hasCreateSuggestion={ false }
+					onChange={ ( updatedValue ) => {
+						updateAttributes(
+							updatedValue,
+							setInsertedBlockAttributes,
+							insertedBlockAttributes
+						);
+						setClientIdWithOpenLinkUI( null );
+					} }
+					onCancel={ () => setClientIdWithOpenLinkUI( null ) }
+				/>
+			)
+		);
+	};
+
 	return (
-		<OffCanvasEditor
-			blocks={ clientIdsTree }
-			parentClientId={ clientId }
-			isExpanded={ true }
-			LeafMoreMenu={ LeafMoreMenu }
-			description={ description }
-		/>
+		<div className="wp-block-navigation__menu-inspector-controls">
+			<PrivateListView
+				blocks={ clientIdsTree }
+				rootClientId={ clientId }
+				isExpanded
+				description={ description }
+				showAppender
+				blockSettingsMenu={ LeafMoreMenu }
+				renderAdditionalBlockUI={ renderLinkUI }
+			/>
+		</div>
 	);
 };
